Removal of Old BR reference - Removal of Leaderboard string input variable
Previously leaderboard string for android had to input as variable of plugin. This has been removed and the leaderboard id is now contained directly in the javascript code, like it is in ios.

The following keys are now used: leaderboardID_droid, leaderboard_ios.

Allows for multiple leaderboards per application for both platforms

This will resolve issue #1